backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.24.0
pydantic==2.10.3
google-generativeai==0.3.2
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0

